+++ build.gradle
@@ -23,8 +23,11 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/Cucumber.xml"
+        //suites "src/test/resources/Cucumber.xml"
         //suites "src/test/resources/TestNG.xml"
+        suites "src/test/resources/SmokeTests.xml"
+        suites "src/test/resources/RegressionTests.xml"
+
         systemProperty "browser", System.getProperty("browser")
         //suites "src/test/resources/NewConfig.xml"
     }
