--- src/test/java/Homework21.java
@@ -1,37 +1,0 @@
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import pagefactory.BasePage;
-import pagefactory.HomePage;
-import pagefactory.LoginPage;
-
-/**
- * @author wiles42
- */
-public class Homework21 extends BaseTest{
-    private static final Logger log = LoggerFactory.getLogger(Homework21.class);
-
-    @Test
-    public void renamePlaylist() {
-        BasePage basePage = new BasePage(driver);
-        LoginPage loginPage = new LoginPage(driver);
-        HomePage homePage = new HomePage(driver);
-
-       //Preconditions
-      String updatedMsg = "Updated playlist \"Wiley Playlist.\"";
-        loginPage.provideEmail("wiley.griffin@testpro.io")
-                .providePassword("KUgY8Y1p")
-                .clickSubmit();
-
-       
-       //Steps
-        homePage.dblClickPlaylist();
-        homePage.inputName("Wiley Playlist");
-
-        //Verification
-        Assert.assertEquals(homePage.successMessage(), updatedMsg);
-
-    }
-
-}
