+++ src/test/resources/RegressionTests.xml
@@ -1,14 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="RegressionTests">
     <parameter name="BaseUrl" value="https://qa.koel.app"/>
 
     <test name="Example Test" preserve-order="false">
         <classes>
             <class name="LoginTests"/>
-            <class name="BaseTest"/>
-           <class name="HomePageTests"/>
             <class name="HomePageTests"/>
             <class name="PlayListTests"/>
         </classes>
