+++ src/test/resources/SmokeTests.xml
@@ -1,15 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="NewConfig">
+<suite name="SmokeTests">
+    <parameter name="BaseUrl" value="https://qa.koel.app"/>
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="HomePageTests">
-                <methods>
-                    <include name="playSong"/>
-                </methods>
-            </class>
+            <class name="LoginTests"/>
+            <class name="HomePageTests"/>
         </classes>
     </test>
 </suite>
